+++ platforms/ios/www/cordova_plugins.js
@@ -1,6 +1,14 @@
 cordova.define('cordova/plugin_list', function(require, exports, module) {
 module.exports = [
   {
+    "id": "com-darryncampbell-cordova-plugin-intent.IntentShim",
+    "file": "plugins/com-darryncampbell-cordova-plugin-intent/www/IntentShim.js",
+    "pluginId": "com-darryncampbell-cordova-plugin-intent",
+    "clobbers": [
+      "intentShim"
+    ]
+  },
+  {
     "id": "cordova-clipboard.Clipboard",
     "file": "plugins/cordova-clipboard/www/clipboard.js",
     "pluginId": "cordova-clipboard",
@@ -295,6 +303,14 @@ module.exports = [
     ]
   },
   {
+    "id": "cordova-plugin-local-notification.LocalNotification",
+    "file": "plugins/cordova-plugin-local-notification/www/local-notification.js",
+    "pluginId": "cordova-plugin-local-notification",
+    "clobbers": [
+      "cordova.plugins.notification.local"
+    ]
+  },
+  {
     "id": "cordova-plugin-media-capture.CaptureAudioOptions",
     "file": "plugins/cordova-plugin-media-capture/www/CaptureAudioOptions.js",
     "pluginId": "cordova-plugin-media-capture",
@@ -414,48 +430,50 @@ module.exports = [
     ]
   },
   {
+    "id": "phonegap-plugin-push.PushNotification",
+    "file": "plugins/phonegap-plugin-push/www/push.js",
+    "pluginId": "phonegap-plugin-push",
+    "clobbers": [
+      "PushNotification"
+    ]
+  },
+  {
     "id": "cordova-sqlite-evcore-extbuild-free.SQLitePlugin",
     "file": "plugins/cordova-sqlite-evcore-extbuild-free/www/SQLitePlugin.js",
     "pluginId": "cordova-sqlite-evcore-extbuild-free",
     "clobbers": [
       "SQLitePlugin"
     ]
-  },
-  {
-    "id": "com-darryncampbell-cordova-plugin-intent.IntentShim",
-    "file": "plugins/com-darryncampbell-cordova-plugin-intent/www/IntentShim.js",
-    "pluginId": "com-darryncampbell-cordova-plugin-intent",
-    "clobbers": [
-      "intentShim"
-    ]
   }
 ];
 module.exports.metadata = 
 // TOP OF METADATA
 {
-  "cordova-clipboard": "1.2.1",
+  "com-darryncampbell-cordova-plugin-intent": "1.1.8",
+  "cordova-android-support-gradle-release": "3.0.1",
+  "cordova-clipboard": "1.3.0",
   "cordova-plugin-badge": "0.8.8",
-  "cordova-plugin-camera": "4.0.3",
-  "cordova-plugin-customurlscheme": "4.3.0",
-  "cordova-plugin-device": "2.0.2",
-  "cordova-plugin-file": "6.0.1",
-  "cordova-plugin-file-opener2": "2.0.19",
+  "cordova-plugin-camera": "4.1.0",
+  "cordova-plugin-customurlscheme": "4.4.0",
+  "cordova-plugin-device": "2.0.3",
+  "cordova-plugin-file": "6.0.2",
+  "cordova-plugin-file-opener2": "2.2.1",
   "cordova-plugin-file-transfer": "1.7.1",
   "cordova-plugin-globalization": "1.11.0",
-  "cordova-plugin-inappbrowser": "3.0.0",
+  "cordova-plugin-inappbrowser": "3.1.0",
   "cordova-plugin-ionic-keyboard": "2.1.3",
-  "cordova-plugin-media-capture": "3.0.2",
-  "cordova-plugin-network-information": "2.0.1",
-  "cordova-plugin-screen-orientation": "3.0.1",
-  "cordova-plugin-splashscreen": "5.0.2",
-  "cordova-plugin-statusbar": "2.4.2",
-  "cordova-plugin-whitelist": "1.3.3",
+  "cordova-plugin-local-notification": "0.9.0-beta.3",
+  "cordova-plugin-media-capture": "3.0.3",
+  "cordova-plugin-network-information": "2.0.2",
+  "cordova-plugin-screen-orientation": "3.0.2",
+  "cordova-plugin-splashscreen": "5.0.3",
+  "cordova-plugin-statusbar": "2.4.3",
+  "cordova-plugin-whitelist": "1.3.4",
   "cordova-plugin-zip": "3.1.0",
-  "cordova-sqlite-storage": "3.2.0",
+  "cordova-sqlite-storage": "3.2.1",
   "nl.kingsquare.cordova.background-audio": "1.0.1",
-  "cordova-sqlite-evcore-extbuild-free": "0.10.1",
-  "com-darryncampbell-cordova-plugin-intent": "1.1.8",
-  "cordova-android-support-gradle-release": "3.0.0"
+  "phonegap-plugin-push": "2.2.3",
+  "cordova-sqlite-evcore-extbuild-free": "0.10.1"
 };
 // BOTTOM OF METADATA
 });