+++ plugins/cordova-plugin-file-opener2/src/android/io/github/pwlin/cordova/plugins/fileopener2/FileOpener2.java
@@ -95,34 +95,36 @@ public class FileOpener2 extends CordovaPlugin {
 		try {
 			CordovaResourceApi resourceApi = webView.getResourceApi();
 			Uri fileUri = resourceApi.remapUri(Uri.parse(fileArg));
-			fileName = this.stripFileProtocol(fileUri.toString());
+			fileName = fileUri.getPath();
 		} catch (Exception e) {
 			fileName = fileArg;
 		}
 		File file = new File(fileName);
 		if (file.exists()) {
 			try {
-				Uri path = Uri.fromFile(file);
-				Intent intent = new Intent(Intent.ACTION_VIEW);
-				if((Build.VERSION.SDK_INT >= 23 && !contentType.equals("application/vnd.android.package-archive")) || ((Build.VERSION.SDK_INT == 24 || Build.VERSION.SDK_INT == 25) && contentType.equals("application/vnd.android.package-archive"))) {
-
-					Context context = cordova.getActivity().getApplicationContext();
-					path = FileProvider.getUriForFile(context, cordova.getActivity().getPackageName() + ".opener.provider", file);
-					intent.setDataAndType(path, contentType);
-					intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-					intent.setFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
-					//intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-
-				 	List<ResolveInfo> infoList = context.getPackageManager().queryIntentActivities(intent, PackageManager.MATCH_DEFAULT_ONLY);
-					for (ResolveInfo resolveInfo : infoList) {
-				    String packageName = resolveInfo.activityInfo.packageName;
-				    context.grantUriPermission(packageName, path, Intent.FLAG_GRANT_WRITE_URI_PERMISSION | Intent.FLAG_GRANT_READ_URI_PERMISSION);
+				Intent intent;
+				if (contentType.equals("application/vnd.android.package-archive")) {
+					// https://stackoverflow.com/questions/9637629/can-we-install-an-apk-from-a-contentprovider/9672282#9672282
+					intent = new Intent(Intent.ACTION_INSTALL_PACKAGE);
+					Uri path;
+					if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) {
+						path = Uri.fromFile(file);
+					} else {
+						Context context = cordova.getActivity().getApplicationContext();
+						path = FileProvider.getUriForFile(context, cordova.getActivity().getPackageName() + ".opener.provider", file);
 					}
-				}
-				else {
 					intent.setDataAndType(path, contentType);
-					intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+					intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION | Intent.FLAG_ACTIVITY_NEW_TASK);
+
+				} else {
+					intent = new Intent(Intent.ACTION_VIEW);
+					Context context = cordova.getActivity().getApplicationContext();
+					Uri path = FileProvider.getUriForFile(context, cordova.getActivity().getPackageName() + ".opener.provider", file);
+					intent.setDataAndType(path, contentType);
+					intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION | Intent.FLAG_GRANT_WRITE_URI_PERMISSION | Intent.FLAG_ACTIVITY_NO_HISTORY);
+
 				}
+
 				/*
 				 * @see
 				 * http://stackoverflow.com/questions/14321376/open-an-activity-from-a-cordovaplugin
@@ -176,13 +178,5 @@ public class FileOpener2 extends CordovaPlugin {
         return appInstalled;
 	}
 
-	private String stripFileProtocol(String uriString) {
-		if (uriString.startsWith("file://")) {
-			uriString = uriString.substring(7);
-		} else if (uriString.startsWith("content://")) {
-			uriString = uriString.substring(10);
-		}
-		return uriString;
-	}
-
 }
+